Fix trunction of flow
A cast to float is require to force the result of the division to be of
type float.

+++ sfm_sf06_i2c.c
@@ -231,7 +231,8 @@ int16_t sfm_sf06_read_measurement_data(float* flow, float* temperature,
         &int_flow, &int_temperature, status_word);
     if (!error) {
         *temperature = int_temperature / 200.0;
-        *flow = (int_flow - _driver.flow_offset) / _driver.flow_scaling_factor;
+        *flow = (float)(int_flow - _driver.flow_offset) /
+                _driver.flow_scaling_factor;
     }
     return error;
 }
